.history/requirements_20250725090348.txt
streamlit==1.35.0
python-dotenv==1.0.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
PyMuPDF==1.22.5
numpy==1.26.4
requests==2.31.0
mistralai==0.0.11
transformers==4.41.1
